Big correctness and vuln audit + fixes (Used Claude Opus 5.5)

+++ src/Core/SourceLocation.h
@@ -8,7 +8,7 @@
 
 typedef struct {
     uint32_t start;
-    uint16_t length;
+    uint32_t length;
     uint16_t fileId;
 } SourceRange;
 
@@ -24,7 +24,7 @@ static inline bool SourceRangeValid(SourceRange r)
 }
 
 #define SRC_INVALID ((SourceRange){0, 0, 0})
-#define SRC_POS(pos, len) ((SourceRange){(uint32_t)(pos), (uint16_t)(len), 0})
+#define SRC_POS(pos, len) ((SourceRange){(uint32_t)(pos), (uint32_t)(len), 0})
 
 typedef struct {
     uint32_t line;   // 1-based
